# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OSM features API",
        "description": "The OSM Features API provides a comprehensive suite of endpoints for accessing OpenStreetMap (OSM) features in GeoJSON format, designed to streamline the integration and utilization of OSM data in applications. By leveraging daily syncs from the OSM database through osm2pgsql imports, the API ensures the delivery of up-to-date and efficiently processed data. Endpoints include retrieving lists and specific details of Places, Administrative Areas, Hiking Routes, and Poles, each identifiable by unique IDs and accompanied by their latest update timestamps. The API allows for fetching single items or lists, providing data in GeoJSON format that encompasses various attributes such as name, class, subclass, and geometry, making it versatile for numerous applications in mapping, navigation, and geographic analysis. This documentation aims to guide users through the API's functionalities, ensuring an understanding of how to make the most of the available resources for their mapping needs. Through concise and practical examples, users can learn how to query the API for specific geographic features and integrate these into their applications, enriching the user experience with detailed and dynamic map-based information.",
        "contact": {
            "email": "info@webmapp.it"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/features/poles/list": {
            "get": {
                "tags": ["API V1"],
                "summary": "List all Poles",
                "description": "Returns a list of Poles with their IDs and updated timestamps. Optionally provide an 'updated_at' parameter to filter poles updated after the specified date. Paginated results are available, with each page containing 100 poles. Use the 'page' parameter to specify the page number to retrieve.",
                "operationId": "listPoles",
                "parameters": [
                    {
                        "name": "updated_at",
                        "in": "query",
                        "description": "Poles updated after this date",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number to retrieve. Each page contains 100 results.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PoleItem"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/features/poles/{id}": {
            "get": {
                "tags": ["API V1"],
                "summary": "Get Pole by ID",
                "description": "Returns a single Pole in GeoJSON format",
                "operationId": "getPoleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Pole ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PoleGeojsonFeature"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pole not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PoleItem": {
                "type": "object",
                "properties": {
                    "osm_id": {
                        "type": "integer",
                        "example": 123
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-03-10T02:00:00Z"
                    }
                }
            },
            "PoleGeojsonFeature": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "Feature"
                    },
                    "properties": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "Sentiero Natura: \"M.te Capre - M.te Palestra - Rifugio M.te Palestra\""
                            },
                            "osm_id": {
                                "type": "integer",
                                "example": 7949696573
                            },
                            "osm_type": {
                                "type": "string",
                                "example": "N"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2021-03-10T02:00:00Z"
                            },
                            "tags": {
                                "type": "json",
                                "example": "{\"name\": \"Sentiero Natura: \\\"M.te Capre - M.te Palestra - Rifugio M.te Palestra\\\"\", \"tourism\": \"information\", \"information\": \"guidepost\"}"
                            },
                            "ref": {
                                "type": "string",
                                "example": "LWN 1"
                            },
                            "ele": {
                                "type": "string",
                                "example": "1000"
                            },
                            "destination": {
                                "type": "string",
                                "example": "Rifugio M.te Palestra"
                            },
                            "support": {
                                "type": "string",
                                "example": "pole"
                            }
                        },
                        "type": "object"
                    },
                    "geometry": {
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "Point"
                            },
                            "coordinates": {
                                "type": "array",
                                "items": {
                                    "type": "number"
                                },
                                "example": [11.8805, 43.4635]
                            }
                        },
                        "type": "object"
                    }
                }
            }
        }
    }
}